+++ src/Redmine/Api/Issue.php
@@ -2,6 +2,7 @@
 
 namespace Redmine\Api;
 
+use Redmine\Client\Client;
 use Redmine\Client\NativeCurlClient;
 use Redmine\Client\Psr18Client;
 use Redmine\Exception;
@@ -81,6 +82,8 @@ class Issue extends AbstractApi
     /**
      * @deprecated v2.9.0 Use fromHttpClient() instead.
      * @see Issue::fromHttpClient()
+     *
+     * @param Client|HttpClient $client
      */
     public function __construct($client/*, bool $privatelyCalled = false*/)
     {